provisioning with puppet

+++ setup/datalog-setup.sh
@@ -1,7 +1,12 @@
 #!/usr/bin/env bash
-cd ~
-wget -O datalog-2.5.tar.gz http://sourceforge.net/projects/datalog/files/datalog/2.5/datalog-2.5.tar.gz/download
+if [ ! -f "datalog-2.5.tar.gz" ]; then 
+	wget -O datalog-2.5.tar.gz http://sourceforge.net/projects/datalog/files/datalog/2.5/datalog-2.5.tar.gz/download
+fi
+
+if [ ! -d "datalog-2.5" ]; then
 tar xzvf datalog-2.5.tar.gz
+fi
+
 cd datalog-2.5
 ./configure
 make
